feat(web): sort albums in modal
